Authors list pagination,CRUD, routing in /config/routes.yaml

+++ 20180802-case/src/Controller/AuthorController.php
@@ -10,6 +10,7 @@ use Symfony\Component\HttpFoundation\Response;
 //move routing to config/routes.yaml
 //use Symfony\Component\Routing\Annotation\Route;
 
+
 /**
  * Route("/author")
  */
@@ -18,13 +19,66 @@ class AuthorController extends Controller
     /**
      * Route("/", name="author_index", methods="GET")
      */
-    public function index(): Response
+    public function index($page): Response
     {
-        $authors = $this->getDoctrine()
-            ->getRepository(Author::class)
-            ->findAll();
+//      simple paginator without preload total count and many other stuff
 
-        return $this->render('author/index.html.twig', ['authors' => $authors]);
+
+        //how many authors display per page, you can change it on yor needs
+        $paginator=5;
+        $repository=$this->getDoctrine()->getRepository(Author::class);
+
+        //many times faster query count without whole data, assuming huge amount
+        $count=$repository->count([]);
+
+
+        $last_page=(integer)ceil($count/$paginator);
+
+//        dump(['total list length'=>$count, 'possible last page'=> $last_page]);
+
+        //process miss formed requests
+
+        if ($page>$last_page) {
+//            dump('reset direct (old, wrong, hacked or forced) page arg');
+            $page=$last_page;
+        }
+
+
+        if ($count > $paginator) {
+
+            $offset=($page-1)*$paginator;
+
+//            dump("Enabling pagination for page $page with offset $offset ");
+
+
+            //findAll changed to limited query https://www.doctrine-project.org/api/orm/2.6/Doctrine/ORM/EntityRepository.html
+
+            //dump($offset);
+            $authors = $repository->findBy([], null, $paginator, $offset);
+
+
+        }
+
+
+        else {
+
+            $authors = $repository->findAll();
+
+
+
+
+
+        }
+
+//        dump(['method'=>__CLASS__, 'page'=>$page, 'displayed list length'=>count($authors)]);
+        return $this->render('author/index.html.twig', [
+            'authors' => $authors,
+            'ion_pager'=>[
+                'per_page'=>$paginator,
+                'total_count'=>$count,
+                'rendered_page_index'=>$page
+            ]
+        ]);
     }
 
     /**
@@ -41,7 +95,7 @@ class AuthorController extends Controller
             $em->persist($author);
             $em->flush();
 
-            return $this->redirectToRoute('author_index');
+            return $this->redirectToRoute('authors');
         }
 
         return $this->render('author/new.html.twig', [
@@ -53,16 +107,21 @@ class AuthorController extends Controller
     /**
      * Route("/{id}", name="author_show", methods="GET")
      */
-    public function show(Author $author): Response
+//    @ParamConverter("id", class="Author")
+// i can't find any ability to convert passed ID to doctrine object via YAML routing (without Route annotations or heavy & not always needed auto convert via app/config/config.yml, see https://symfony.com/doc/current/bundles/SensioFrameworkExtraBundle/annotations/converters.html ), so i must convert passed id inside action
+    public function show(Author $author, $id): Response
     {
+//        dump([$this->getDoctrine()->getRepository(Author::class)->findOneBy(['id'=>$id]), $id]);
+        $author=$this->getDoctrine()->getRepository(Author::class)->findOneBy(['id'=>$id]);
         return $this->render('author/show.html.twig', ['author' => $author]);
     }
 
     /**
      * Route("/{id}/edit", name="author_edit", methods="GET|POST")
      */
-    public function edit(Request $request, Author $author): Response
+    public function edit(Request $request, Author $author, $id): Response
     {
+        $author=$this->getDoctrine()->getRepository(Author::class)->findOneBy(['id'=>$id]);
         $form = $this->createForm(AuthorType::class, $author);
         $form->handleRequest($request);
 
@@ -81,14 +140,20 @@ class AuthorController extends Controller
     /**
      * Route("/{id}", name="author_delete", methods="DELETE")
      */
-    public function delete(Request $request, Author $author): Response
+    public function delete(Request $request, Author $author, $id): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$author->getId(), $request->request->get('_token'))) {
+        $author=$this->getDoctrine()->getRepository(Author::class)->findOneBy(['id'=>$id]);
+
+//        dd($author);
+
+        if ($author && $this->isCsrfTokenValid('delete'.$author->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($author);
             $em->flush();
         }
 
-        return $this->redirectToRoute('author_index');
+
+
+        return $this->redirectToRoute('authors');
     }
 }
